+++ src/main/java/org/betonquest/betonquest/api/QuestCompassTargetChangeEvent.java
@@ -1,6 +1,8 @@
 package org.betonquest.betonquest.api;
 
+import lombok.Getter;
 import org.bukkit.Location;
+import org.bukkit.entity.Player;
 import org.bukkit.event.Cancellable;
 import org.bukkit.event.Event;
 import org.bukkit.event.HandlerList;
@@ -13,20 +15,20 @@ public class QuestCompassTargetChangeEvent extends Event implements Cancellable 
 
     private static final HandlerList HANDLERS = new HandlerList();
 
+    @Getter
+    private final Player player;
+    @Getter
     private final Location location;
     private boolean cancelled;
 
-    public QuestCompassTargetChangeEvent(final Location location) {
+    public QuestCompassTargetChangeEvent(final Player player, final Location location) {
         super();
+        this.player = player;
         this.location = location;
     }
 
     public static HandlerList getHandlerList() {
         return HANDLERS;
-    }
-
-    public Location getLocation() {
-        return location;
     }
 
     @Override